requirements.txt
essentials
click
flake8
black
pytest
pytest-cov
python-dotenv
mypy
